refacto: define important kernels in separate files
